refactor(notebook): remove shared compatibility shims
