+++ apps/elements/tsconfig.json
@@ -19,8 +19,6 @@
       "@/components/isolated/*": ["../../src/components/isolated/*"],
       "@/components/notebook": ["../../src/components/notebook/index.ts"],
       "@/components/notebook/*": ["../../src/components/notebook/*"],
-      "@/components/notebook-shell": ["../../src/components/notebook-shell/index.ts"],
-      "@/components/notebook-shell/*": ["../../src/components/notebook-shell/*"],
       "@/components/notebook-rail": ["../../src/components/notebook-rail/index.ts"],
       "@/components/notebook-rail/*": ["../../src/components/notebook-rail/*"],
       "@/components/outputs/*": ["../../src/components/outputs/*"],
