--- apps/notebook/src/components/CondaDependencyHeader.tsx
@@ -1,4 +1,0 @@
-export {
-  CondaDependencyPanel as CondaDependencyHeader,
-  type CondaDependencyPanelProps as CondaDependencyHeaderProps,
-} from "@/components/environment";
