--- apps/notebook/src/components/DaemonStatusBanner.tsx
@@ -1,5 +1,0 @@
-export {
-  DaemonStatusBanner,
-  type DaemonStatus,
-  type DaemonStatusBannerProps,
-} from "@/components/notebook/DaemonStatusBanner";
