--- apps/notebook/src/components/DebugBanner.tsx
@@ -1,1 +1,0 @@
-export { DebugBanner, type DebugBannerProps } from "@/components/notebook/DebugBanner";
