--- apps/notebook/src/components/DenoDependencyHeader.tsx
@@ -1,5 +1,0 @@
-export {
-  DenoDependencyPanel as DenoDependencyHeader,
-  type DenoConfigInfo,
-  type DenoDependencyPanelProps as DenoDependencyHeaderProps,
-} from "@/components/environment";
