--- apps/notebook/src/components/DependencyHeader.tsx
@@ -1,4 +1,0 @@
-export {
-  UvDependencyPanel as DependencyHeader,
-  type UvDependencyPanelProps as DependencyHeaderProps,
-} from "@/components/environment";
