--- apps/notebook/src/components/EnvBuildDecisionDialog.tsx
@@ -1,5 +1,0 @@
-export {
-  EnvBuildDecisionDialog,
-  extractCondaEnvCreateCommand,
-  type EnvBuildDecisionDialogProps,
-} from "@/components/notebook/EnvBuildDecisionDialog";
