--- apps/notebook/src/components/GlobalFindBar.tsx
@@ -1,1 +1,0 @@
-export { GlobalFindBar, type GlobalFindBarProps } from "@/components/search";
