--- apps/notebook/src/components/KernelLaunchErrorBanner.tsx
@@ -1,5 +1,0 @@
-export {
-  KernelLaunchErrorBanner,
-  shouldShowKernelLaunchErrorBanner,
-  type KernelLaunchErrorBannerProps,
-} from "@/components/notebook/KernelLaunchErrorBanner";
