--- apps/notebook/src/components/PackageSpecList.tsx
@@ -1,6 +1,0 @@
-export {
-  PackageSpecList,
-  parsePackageSpec,
-  type PackageSpecListProps,
-  type PackageSpecTone,
-} from "@/components/environment";
