--- apps/notebook/src/components/RuntimeDecisionDialog.tsx
@@ -1,4 +1,0 @@
-export {
-  RuntimeDecisionDialog,
-  type RuntimeDecisionDialogProps,
-} from "@/components/notebook/RuntimeDecisionDialog";
