--- apps/notebook/src/components/TrustDialog.tsx
@@ -1,1 +1,0 @@
-export { TrustDialog, type TrustDialogProps } from "@/components/notebook/TrustDialog";
