--- apps/notebook/src/components/UntrustedBanner.tsx
@@ -1,1 +1,0 @@
-export { UntrustedBanner, type UntrustedBannerProps } from "@/components/notebook/UntrustedBanner";
