--- apps/notebook/src/components/icons.tsx
@@ -1,8 +1,0 @@
-export {
-  CondaIcon,
-  DenoIcon,
-  PixiIcon,
-  PythonIcon,
-  UvIcon,
-  type EnvironmentIconProps,
-} from "@/components/environment";
