--- apps/notebook/src/components/runtime-surface-types.ts
@@ -1,7 +1,0 @@
-export type {
-  EnvSyncState,
-  PyProjectDeps,
-  PyProjectInfo,
-  TrustInfo,
-  TyposquatWarning,
-} from "@/components/notebook/runtime-surface-types";
