--- src/components/notebook-shell/NotebookCellList.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookCellList";
