--- src/components/notebook-shell/NotebookCommandToolbar.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookCommandToolbar";
