--- src/components/notebook-shell/NotebookDocumentHeader.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookDocumentHeader";
