--- src/components/notebook-shell/NotebookDocumentRail.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookDocumentRail";
