--- src/components/notebook-shell/NotebookDocumentShell.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookDocumentShell";
