--- src/components/notebook-shell/NotebookEditModeButton.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookEditModeButton";
