--- src/components/notebook-shell/NotebookEnvironmentSummary.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookEnvironmentSummary";
