--- src/components/notebook-shell/NotebookIdentity.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookIdentity";
