--- src/components/notebook-shell/NotebookPackageSummaryPanel.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookPackageSummaryPanel";
