--- src/components/notebook-shell/NotebookPresenceStatus.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookPresenceStatus";
