--- src/components/notebook-shell/NotebookReadOnlyView.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookReadOnlyView";
