--- src/components/notebook-shell/NotebookToolbarFrame.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookToolbarFrame";
