--- src/components/notebook-shell/NotebookToolbarIdentity.tsx
@@ -1,1 +1,0 @@
-export * from "../notebook/NotebookToolbarIdentity";
