--- src/components/notebook-shell/actor-labels.ts
@@ -1,1 +1,0 @@
-export * from "../notebook/actor-labels";
