--- src/components/notebook-shell/actor-projection.ts
@@ -1,1 +1,0 @@
-export * from "../notebook/actor-projection";
