--- src/components/notebook-shell/capabilities.ts
@@ -1,1 +1,0 @@
-export * from "../notebook/capabilities";
