--- src/components/notebook-shell/environment-surface.ts
@@ -1,1 +1,0 @@
-export * from "../notebook/environment-surface";
