--- src/components/notebook-shell/index.ts
@@ -1,1 +1,0 @@
-export * from "../notebook";
