--- src/components/notebook-shell/interaction-mode.ts
@@ -1,1 +1,0 @@
-export * from "../notebook/interaction-mode";
