--- src/components/notebook-shell/outline-navigation.ts
@@ -1,1 +1,0 @@
-export * from "../notebook/outline-navigation";
