--- src/components/notebook-shell/view-model.ts
@@ -1,1 +1,0 @@
-export * from "../notebook/view-model";
